Lay the foundation.
Add a TravisCI configuration file so we can get this building on Travis.

Update the default README using a script to contain the godoc.

Add the script that updates the README.

Create the foundation of what a payment is.

Create an interface for storing payments and retrieving payments.

Create an in-memory implementation for storing payments.